requirements.txt
# Common requirements
build>=1.2.1
setuptools~=67.8.0
wheel~=0.38.4
requests>=2.29.0
python-decouple>=3.8
flask==3.1.0
flask-restx==1.3.0
typing-extensions~=4.12.2
pyjwt>=2.8.0
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
# SQLAlchemy==1.4.47
# Flask-SQLAlchemy==3.0.3
cryptography>=41.0.7


# Tests requirements
requests-mock>=1.12.1
pytest>=7.4.0
coverage>=6.4.4
flake8>=4.0.1
